Returned getters and setters to custom type Component
These getters and setters are styled not in the way that JavaScript has
readily available but in the way that one might have them styled in C++ or
Java. I did it this way because I don't know how a subtype of Component
would otherwise be able to call the getters and setters (with the call
function); custom type Button needs those getters and setters.